pyproject.toml
[project]
name = "data89-evaluation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "arviz>=0.22.0",
    "certifi>=2025.11.12",
    "cmdstanpy>=1.3.0",
    "httpstan>=4.13.0",
    "ipywidgets>=8.1.8",
    "joblib>=1.5.2",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "plotnine>=0.15.2",
    "pyhere>=1.0.3",
    "pyjanitor>=0.32.3",
    "scipy>=1.16.3",
    "statsmodels>=0.14.6",
]

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"
